@@ -1,5 +1,5 @@
 import './App.css';
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Home from './pages/Home/Home.jsx';
 import Edit from './pages/Edit/Edit.jsx';
 import EmailOption from './pages/EmailOption/EmailOption.jsx';
@@ -11,17 +11,17 @@ function App() {
   return (
     <Router>
       <div className="App">
-        <Switch>
-          <Route path="/" exact component={Home} />
-          <Route path="/edit" component={Edit} />
-          <Route path="/email-option" component={EmailOption} />
-          <Route path="/profile" component={Profile} />
-          <Route path="/signin" component={SignIn} />
-          <Route path="/signup" component={SignUp} />
-          
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/edit" element={<Edit />} />
+          <Route path="/email-option" element={<EmailOption />} />
+          <Route path="/profile" element={<Profile />} />
+          <Route path="/signin" element={<SignIn />} />
+          <Route path="/signup" element={<SignUp />} />
+
           {/* Optional: Route for 404 Not Found */}
-          <Route render={() => <h2>404 Not Found</h2>} />
-        </Switch>
+          <Route path="*" element={<h2>404 Not Found</h2>} />
+        </Routes>
       </div>
     </Router>
   );
