@@ -3,7 +3,59 @@ import "./profile.scss"
 
 const Profile = () => {
   return (
-    <div>Profile</div>
+   <>
+   <div className="w-[1440px] h-[1022px] relative bg-white">
+  <div className="w-[1440px] h-[1024px] left-0 top-[-2px] absolute">
+    <div className="w-[1440px] h-[1024px] left-0 top-0 absolute bg-neutral-50" />
+    <div className="w-[1440px] h-[216.95px] left-0 top-0 absolute bg-zinc-100" />
+    <div className="w-[1440px] h-[3.25px] left-0 top-[216.95px] absolute bg-neutral-200" />
+    <div className="w-[704px] h-0.5 left-[593px] top-[320px] absolute origin-top-left rotate-90 bg-neutral-200" />
+    <div className="w-[1440px] h-[1024px] left-0 top-0 absolute bg-zinc-300" />
+  </div>
+  <div className="w-[293px] h-[319.73px] left-[183px] top-[308px] absolute">
+    <img className="w-[229.75px] h-[210.70px] left-[31.03px] top-0 absolute rounded-full border-4 border-stone-400" src="https://via.placeholder.com/230x211" />
+    <div className="left-0 top-[219px] absolute text-black text-[33px] font-normal font-['Agbalumo']">Kushagra   Agrawal</div>
+    <div className="w-[268.84px] h-[34.73px] left-[21px] top-[285px] absolute text-neutral-500 text-2xl font-normal font-['Acme']">Undergraduate Student</div>
+  </div>
+  <div className="w-[416.85px] h-[80.15px] left-[121px] top-[665px] absolute">
+    <div className="w-[63px] left-0 top-[56.08px] absolute text-neutral-500 text-base font-normal font-['Actor']">Domains</div>
+    <div className="left-[157px] top-[56px] absolute text-neutral-500 text-base font-normal font-['Actor']">Requests</div>
+    <div className="w-[95px] left-[321.85px] top-[56.08px] absolute text-neutral-500 text-base font-normal font-['Actor']">Emails Added</div>
+    <div className="left-[23px] top-0 absolute text-black text-[33.01px] font-normal font-['Abel']">4</div>
+    <div className="w-[50px] left-[161.61px] top-[0.38px] absolute text-black text-[33.01px] font-normal font-['Abel']">238</div>
+    <div className="left-[356px] top-0 absolute text-black text-[33.01px] font-normal font-['Abel']">2</div>
+    <div className="w-[1.38px] h-[68.77px] left-[95.59px] top-[11.38px] absolute origin-top-left rotate-180 bg-neutral-500" />
+    <div className="w-[1.38px] h-[68.77px] left-[273.71px] top-[10px] absolute origin-top-left rotate-180 bg-neutral-500" />
+  </div>
+  <div className="w-[333px] h-[62.53px] left-[152px] top-[807px] absolute">
+    <div className="w-[333px] h-[62.53px] left-0 top-0 absolute bg-neutral-500 rounded-[10px]" />
+    <div className="w-[223.13px] h-[38.77px] left-[62.30px] top-[11.25px] absolute text-white text-2xl font-normal font-['ADLaM Display']">kush@gmail.com</div>
+  </div>
+  <div className="w-[526px] left-[718px] top-[902px] absolute text-neutral-500 text-sm font-light font-['Helvetica']">I’m web designer, I work in programs like figma, adobe photoshop, adobe illustrator</div>
+  <div className="w-[714px] h-[686px] left-[627px] top-[294px] absolute">
+    <div className="w-[398.51px] h-[39.69px] left-0 top-[153.07px] absolute rounded-[7px] border-2 border-neutral-200" />
+    <div className="w-[293.35px] h-[39.69px] left-[420.65px] top-[153.07px] absolute rounded-[7px] border-2 border-neutral-200" />
+    <div className="w-[714px] h-[39.69px] left-0 top-[259.66px] absolute rounded-[7px] border-2 border-neutral-200" />
+    <div className="w-[714px] h-[113.39px] left-0 top-[572.61px] absolute rounded-[7px] border-2 border-neutral-200" />
+    <div className="w-[714px] h-[39.69px] left-0 top-[366.24px] absolute rounded-[7px] border-2 border-neutral-200" />
+    <div className="w-[47.60px] h-[21.54px] left-0 top-[327.69px] absolute text-black text-sm font-normal font-['Helvetica']">EMAIL</div>
+    <div className="w-[80.81px] h-[19.28px] left-0 top-[221.11px] absolute text-black text-sm font-normal font-['ABeeZee']">User Name</div>
+    <div className="w-[94.09px] h-[21.54px] left-0 top-[114.52px] absolute text-black text-sm font-normal font-['Helvetica']">FIRST NAME</div>
+    <div className="w-[116.23px] h-[28.35px] left-0 top-0 absolute text-black text-lg font-bold font-['Helvetica']">BASIC INFO</div>
+    <div className="w-[106.27px] h-[28.35px] left-0 top-[451.29px] absolute text-black text-lg font-bold font-['Helvetica']">ABOUT ME</div>
+    <div className="w-[87.45px] h-[21.54px] left-[420.65px] top-[114.52px] absolute text-black text-sm font-normal font-['Helvetica']">LAST NAME</div>
+    <div className="w-[714px] h-[2.27px] left-0 top-[73.70px] absolute bg-neutral-200" />
+    <div className="w-[714px] h-[2.27px] left-0 top-[524.99px] absolute bg-neutral-200" />
+  </div>
+  <div className="left-[130px] top-[79px] absolute text-black text-[115px] font-normal font-['ADLaM Display']">Profile</div>
+  <div className="w-[373px] h-[150px] left-[792px] top-[79px] absolute bg-neutral-400 rounded-full" />
+  <div className="left-[834px] top-[118px] absolute text-neutral-900 text-5xl font-normal font-['Agbalumo']">Mail Mastery</div>
+  <div className="w-[195px] h-[46px] left-[219px] top-[912px] absolute bg-zinc-500 rounded-[40px]" />
+  <div className="left-[246px] top-[902px] absolute text-white text-[38px] font-normal font-['Agbalumo']">Sign out</div>
+  <div className="w-[118px] h-[39px] left-[621px] top-[229px] absolute bg-stone-300 rounded-[25px]" />
+  <div className="w-[61px] left-[632px] top-[229px] absolute text-black text-[25px] font-normal font-['Akatab']">Edit</div>
+  <img className="w-[41.40px] h-[27px] left-[688px] top-[233px] absolute" src="https://via.placeholder.com/41x27" />
+</div></>
   )
 }
 
